Remove innter astype

+++ chainer/gradient_check.py
@@ -289,10 +289,8 @@ def check_backward(func, x_data, y_grad, params=(),
         for skip, cx, data in six.moves.zip(no_grads, casted_xs, casted_data):
             if skip:
                 continue
-            # Inner astype is required to execute __mul__ with the given
-            # type, and outer astype is require to store data with the given
-            # type
-            data = (one * data.astype(cx.data.dtype)).astype(cx.data.dtype)
+            # astype is require to store data with the given type
+            data = (one * data).astype(data.dtype)
             if numpy.isscalar(data):
                 data = xp.array(data)
             cx.data = data
@@ -301,7 +299,7 @@ def check_backward(func, x_data, y_grad, params=(),
                 param_dtype = dtype
             else:
                 param_dtype = param.dtype
-            param.data = (one * data.astype(param_dtype)).astype(param_dtype)
+            param.data = (one * data).astype(param_dtype)
         ys = func(*casted_xs)
         ys = _as_tuple(ys)
         ys_data = tuple(y.data for y in ys)
